requirement.txt
pyvisa
matplotlib
numpy
requests
Pillow
screeninfo
pyserial
pandas
scipy
